test: fix verfication events
expectEmit expect events on the next call.
With the lib HuffDeployer, there is a first call
to create the HuffConfig. It then call the latter
for configuration. recordLogs is necessary to
capture events.

Precedent version of huff is needed for test to pass:
huffup -v nightly-94c34e402a46365fee29863aa08558af178c2b94